refactor dotcom-build-webpack-config into a webpack plugin
