--- packages/dotcom-build-webpack-config/src/entities/hooks.ts
@@ -1,10 +1,0 @@
-export enum hooks {
-  BABEL_CONFIG = 'babelConfig',
-  BABEL_PRESET_ENV_OPTIONS = 'babelConfig::preset::env::options',
-  WEBPACK_CLEAN_PLUGIN_OPTIONS = 'webpackConfig::plugins::cleanWebpackPlugin::options',
-  WEBPACK_JS_RULE = 'webpackConfig::jsRule',
-  WEBPACK_MANIFEST_PLUGIN_OPTIONS = 'webpackConfig::plugins::manifestPlugin::options',
-  WEBPACK_GZIP_COMPRESSION_PLUGIN_OPTIONS = 'webpackConfig::plugins::gzipCompressionPlugin::options',
-  WEBPACK_BROTLI_COMPRESSION_PLUGIN_OPTIONS = 'webpackConfig::plugins:brotliCompressionPlugin::options',
-  WEBPACK_CONFIG = 'webpackConfig'
-}
