--- packages/dotcom-build-webpack-config/src/types/HandlerArgs.d.ts
@@ -1,6 +1,0 @@
-import { ConfigContext } from '../entities/ConfigContext'
-import * as pluggable from '@financial-times/dotcom-page-kit-pluggable'
-
-export interface HandlerArgs extends pluggable.HandlerArgs {
-  context: ConfigContext
-}
